% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-features.R
\name{add_features_flr}
\alias{add_features_flr}
\title{Add some possible clustering features to FLR output}
\usage{
add_features_flr(data_mod, data_ts, n_years_slope = 5)
}
\arguments{
\item{data_mod}{Model output data.}

\item{data_ts}{Model output time series data.}

\item{n_years_slope}{Number of years for the slope calculation on F and SSB.}
}
\description{
Add some possible clustering features to FLR output
}
\examples{
add_features_flr(haddock_mod, haddock_ts, n_years_slope = 5)
}
