% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haddock_mod}
\alias{haddock_mod}
\alias{haddock_ts}
\alias{cod_ts}
\alias{cod_mod}
\alias{sims3_ts}
\alias{sims3_mod}
\alias{simg3_ts}
\alias{simg3_mod}
\alias{pollock_ts}
\alias{pollock_mod}
\alias{redgrouper_ts}
\alias{redgrouper_mod}
\title{Example model and time series data}
\format{Data frames}
\usage{
haddock_mod

haddock_ts

cod_ts

cod_mod

sims3_ts

sims3_mod

simg3_ts

simg3_mod

pollock_ts

pollock_mod

redgrouper_ts

redgrouper_mod
}
\description{
Example model and time series data
}
\keyword{datasets}
