% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_simple}
\alias{ensemble_simple}
\title{Ensemble multiple input values}
\usage{
ensemble_simple(value, weights = NULL)
}
\arguments{
\item{weights}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}

\item{.data}{a code{data.frame} with columns holding the data to
ensemble over, cluster ids, and optional weights}

\item{method}{one of: simple, cluster, 2stage (see details)}

\item{cluster_id}{a vector of cluster ids defining the cluster
membership of the value}
}
\value{
vector of numeric values
}
\description{
Creates an ensemble of multiple input values based on provided
weightss
}
\details{
not yet!
}
\examples{

# get data
data(haddock_mod)
df <-
  haddock_mod \%>\%
  select(fo_median, ffmsy_median, bbmsy_median)

# perform simple clustering on a few quantities
m <-
  df \%>\% kmeans(centers = 2)

# a plot
factoextra::fviz_cluster(m, data = df, axis = c(1,2))

# calculate weighting?
df$cluster <- m$cluster

# do a simple ensemble - no weighting
df \%>\%
  mutate(
    ens_simple = ensemble_simple(ffmsy_median)
  )

# do a simple cluster ensemble - no weighting
df \%>\%
  group_by(cluster) \%>\%
  mutate(
    ens_cluster = ensemble_simple(ffmsy_median)
  )

}
