% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits2data_mod.R
\name{fits2data_mod}
\alias{fits2data_mod}
\title{Get model level data}
\usage{
fits2data_mod(fits, fo, ffmsy, bbmsy)
}
\arguments{
\item{fits}{a list containing the elements}

\item{fo}{a matrix containing simulated values for each model}

\item{ffmsy}{a matrix containing simulated values for each model}

\item{bbmsy}{a matrix containing simulated values for each model}
}
\value{
A data.frame object
}
\description{
Create a time series dataset from a list of FLStocks and FL fit objects
}
