% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits2data_ts.R
\name{fits2data_ts}
\alias{fits2data_ts}
\title{Create time series dataset}
\usage{
fits2data_ts(fits)
}
\arguments{
\item{fits}{a list containing the elements}
}
\value{
A data.frame object
}
\description{
Create a time series dataset from a list of FLStocks and FL fit objects
}
