% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters}
\usage{
plot_clusters(model, data = NULL, colour_vector = NULL,
  colour_label = "model", ...)
}
\arguments{
\item{model}{An object of class "partition" created by the functions
\code{\link[cluster:pam]{cluster::pam()}} or \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{data}{The original data frame used in the clustering.}

\item{colour_vector}{A vector of character or factors to colour the points
by.}

\item{colour_label}{A label for the colour legend.}

\item{...}{Other arguments to pass2 \code{\link[factoextra:fviz_cluster]{factoextra::fviz_cluster()}}.}
}
\value{
A ggplot2 object.
}
\description{
Plot clusters
}
\examples{
df <- haddock_mod \%>\%
  dplyr::select(fo_median, ffmsy_median, bbmsy_median) \%>\%
  scale()
kmeans(df, centers = 2L)
plot_clusters(m, data = df, colour_vector = haddock_mod$fmodel,
  colour_label = "F model")
}
