% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-som.R
\name{plot_som_individual}
\alias{plot_som_individual}
\alias{plot_som}
\title{Plot self-organizing map}
\usage{
plot_som_individual(model, data, val, colour_vector = NULL,
  colour_label = "model")

plot_som(model, data, colour_vector = NULL, colour_label = "model",
  ...)
}
\arguments{
\item{model}{An object as output from \code{\link[kohonen:som]{kohonen::som()}}.}

\item{data}{The original data frame used in the clustering.}

\item{val}{name of value to plot that matches the variable in the data column}

\item{colour_vector}{A vector of character or factors to colour the points
by.}

\item{colour_label}{A label for the colour legend.}

\item{...}{Other arguments to past to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.}
}
\value{
A ggplot2 object.
}
\description{
Plot self-organizing map
}
\examples{
df <- haddock_mod \%>\%
  dplyr::select(fo_median, ffmsy_median, bbmsy_median) \%>\%
  scale()

# Create the SOM Grid - you generally have to specify the size of the
# training grid prior to training the SOM. Hexagonal and Circular
# topologies are possible
# dimensions chosen based on aiming for 5-10 assessments per class
som_grid <- kohonen::somgrid(xdim = 3, ydim = 2, topo = "hexagonal")

# Finally, train the SOM, options for the number of iterations,
# the learning rates, and the neighbourhood are available
som_model <- kohonen::som(df,
  grid = som_grid,
  rlen = 500,
  alpha = c(0.05, 0.01),
  keep.data = TRUE
)

plot_som_individual(som_model,
  data = df, val = "ffmsy_median",
  colour_vector = haddock_mod$fmodel,
  colour_label = "F model"
)

plot_som(som_model,
  data = df, colour_vector = haddock_mod$fmodel,
  colour_label = "F model",
  nrow = 2
)
}
