% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensembleMethods}
\alias{ensembleMethods}
\alias{ensemble_simple}
\alias{ensemble_2stage_weights}
\alias{ensemble_2stage}
\title{Ensemble multiple input values}
\usage{
ensemble_simple(value, weights = NULL)

ensemble_2stage_weights(cluster, weights = NULL)

ensemble_2stage(value, cluster, weights = NULL)
}
\arguments{
\item{value}{a vector holding the values to ensemble over}

\item{weights}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}

\item{cluster}{a vector of cluster ids defining the cluster
membership of the value}
}
\value{
vector of numeric values
}
\description{
Creates an ensemble of multiple input values based on provided
weightss
}
\details{
not yet!
}
\examples{

library(dplyr)

# get data
data(haddock_mod)
df <-
  haddock_mod \%>\%
  select(fo_median, ffmsy_median, bbmsy_median)

# perform simple clustering on a few quantities
m <-
  df \%>\% kmeans(centers = 2)

# a plot
factoextra::fviz_cluster(m, data = df, axis = c(1,2))

# calculate weighting?
df$cluster <- m$cluster

# do a simple ensemble - no weighting
df \%>\%
  mutate(
    ens_simple = ensemble_simple(ffmsy_median)
  )

# do a simple cluster ensemble - no weighting
df \%>\%
  group_by(cluster) \%>\%
  mutate(
    ens_cluster = ensemble_simple(ffmsy_median)
  )


# fake skill measure for now
df$cgcv <- runif(nrow(df),0,1)

df \%>\%
  mutate(
    ens_cluster_wts  = ensemble_2stage_weights(cluster, cgcv)
  ) \%>\%
  mutate(
    ens_cluster = ensemble_simple(ffmsy_median, ens_cluster_wts)
  )

}
